Add adversarial self-play environment
